+++ CHANGES.rst
@@ -8,6 +8,7 @@ Dev
 - Add source maps support for SASS/SCSS
 - Add source maps support for LESS
 - Add source maps support for CoffeeScript
+- Add source maps support for Stylus
 
 
 1.0.1